Add typings for create_passive_sell_offers.
